Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
openai
python-dotenv
scikit-learn
joblib
numpy
pandas
